Add test scripts

+++ mongodb/delete.sh
@@ -4,8 +4,6 @@ currentDir="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 cd $currentDir
 . functions.sh
 
-confirm || exit 1
-
 optName=$1
 optVersion=$2
 optPort=$3
@@ -17,5 +15,5 @@ sleep 1
 set -eu
 exitIfNotExistDir $dir/datadir/$optName
 exitIfRunningPort $optPort
-rm -fr $dir/datadir/$optName
+[ -d "$dir/datadir/$optName" ] && rm -fr $dir/datadir/$optName
 echo MongoDB Successfully deleted. $optName $optVersion $optPort
